chore(worker-sessions): restore package file-count baseline
